refactor(api): split delete and getUserById from user service
